public re-usable Main
